fix: rename project module db -> database
